// src/app/interfaces/okm-consentimientoam.interface.ts
export interface OKM_ConsentimientoAm {
  name?: string;
  parent?: string;
  parentfield?: string;
  parenttype?: string;
  tipcon_codigo?: string;
  conam_fecha?: string;
  conam_firma?: string;
  conam_documento?: string;
  conam_contestigo?: string;
  conam_nomtestigo?: string;
  conam_idtestigo?: string;
  conam_firmatestigo?: string;
}